# feat: add token reduction for text optimization