--- a/kreuzberg/_token_reduction/__init__.py
+++ b/kreuzberg/_token_reduction/__init__.py
@@ -1,0 +1,11 @@
+from __future__ import annotations
+
+from kreuzberg._token_reduction._reducer import ReductionStats, get_reduction_stats, reduce_tokens
+from kreuzberg._token_reduction._stopwords import StopwordsManager
+
+__all__ = [
+    "ReductionStats",
+    "StopwordsManager",
+    "get_reduction_stats",
+    "reduce_tokens",
+]