… .github/workflows/ci.yaml
… .pre-commit-config.yaml
… ATTRIBUTIONS.md
… benchmarks/token_reduction_compression_benchmark.py
… docker-logs/docker-info.txt
… docker-logs/docker-version.txt
… docs/api-reference/types.md
… docs/user-guide/index.md
… docs/user-guide/token-reduction.md
… kreuzberg/_token_reduction/stop_words.json
… kreuzberg/_token_reduction/stopwords/af_stopwords.json
… kreuzberg/_token_reduction/stopwords/ar_stopwords.json
… kreuzberg/_token_reduction/stopwords/bg_stopwords.json
… kreuzberg/_token_reduction/stopwords/bn_stopwords.json
… kreuzberg/_token_reduction/stopwords/br_stopwords.json
… kreuzberg/_token_reduction/stopwords/ca_stopwords.json
… kreuzberg/_token_reduction/stopwords/cs_stopwords.json
… kreuzberg/_token_reduction/stopwords/da_stopwords.json
… kreuzberg/_token_reduction/stopwords/de_stopwords.json
… kreuzberg/_token_reduction/stopwords/el_stopwords.json
… kreuzberg/_token_reduction/stopwords/en_stopwords.json
… kreuzberg/_token_reduction/stopwords/eo_stopwords.json
… kreuzberg/_token_reduction/stopwords/es_stopwords.json
… kreuzberg/_token_reduction/stopwords/et_stopwords.json
… kreuzberg/_token_reduction/stopwords/eu_stopwords.json
… kreuzberg/_token_reduction/stopwords/fa_stopwords.json
… kreuzberg/_token_reduction/stopwords/fi_stopwords.json
… kreuzberg/_token_reduction/stopwords/fr_stopwords.json
… kreuzberg/_token_reduction/stopwords/ga_stopwords.json
… kreuzberg/_token_reduction/stopwords/gl_stopwords.json