## Summary
- Implements comprehensive token reduction functionality to optimize extracted text for LLMs and storage systems
- Supports two reduction modes: light (formatting only) and moderate (+ stopword removal)
- Adds support for 64+ languages with intelligent stopword detection

## Features
- **Light mode**: Formatting optimization only (~10% reduction, 0.03ms)
- **Moderate mode**: Formatting + stopword removal (~35% reduction, 0.23ms)
- **Language support**: 64+ languages with built-in stopwords
- **Markdown preservation**: Maintains document structure during reduction
- **Custom stopwords**: Domain-specific stopword support per language
- **TypedDict statistics**: Type-safe reduction metrics

## Implementation Details
- Pure functional API design with kwargs for 2+ parameters
- Pre-compiled regex patterns for optimal performance
- LRU caching and Ref pattern for efficient stopwords management
- Security validation (10MB text limit, language code validation)
- Comprehensive test suite with 44 tests covering edge cases
- Full user documentation with performance benchmarks

## Performance Benchmarks
Based on comprehensive testing across different text types:

**Light Mode**
- Character reduction: 10.1% average
- Token reduction: 0% (preserves all words)
- Processing time: 0.03ms average

**Moderate Mode**
- Character reduction: 35.3% average (11.4% - 62.3% range)
- Token reduction: 33.7% average
- Processing time: 0.23ms average

## Testing
- 44 comprehensive tests covering all functionality
- Edge case coverage including unicode, mixed languages, markdown nesting
- Integration tests with full extraction pipeline
- Security validation tests
- 100% type safety (mypy clean)

## Documentation
- User guide with examples and use cases
- API reference documentation
- Performance benchmarks included
- Attribution for stopwords data source

## Files Changed
- Core implementation in `kreuzberg/_token_reduction/`
- Tests in `tests/features/` and `tests/integration/`
- Documentation in `docs/user-guide/`
- Benchmark suite in `benchmarks/`
- Public API exports in `kreuzberg/__init__.py`
- Integration with extraction pipeline

## Breaking Changes
None - this is a new opt-in feature with default mode="off"